Refactor the model manager
