+++ src/cpp/lemonade_router/model_manager.cpp
@@ -47,160 +47,6 @@ static bool contains_ignore_case(const std::string& str, const std::string& subs
 }
 
 // Structure to hold identified GGUF files
-struct GGUFFiles {
-    std::map<std::string, std::string> core_files;  // {"variant": "file.gguf", "mmproj": "file.mmproj"}
-    std::vector<std::string> sharded_files;         // Additional shard files
-};
-
-// Identifies GGUF model files matching the variant (Python equivalent of identify_gguf_models)
-static GGUFFiles identify_gguf_models(
-    const std::string& checkpoint,
-    const std::string& variant,
-    const std::string& mmproj,
-    const std::vector<std::string>& repo_files
-) {
-    const std::string hint = R"(
-    The CHECKPOINT:VARIANT scheme is used to specify model files in Hugging Face repositories.
-
-    The VARIANT format can be one of several types:
-    0. wildcard (*): download all .gguf files in the repo
-    1. Full filename: exact file to download
-    2. None/empty: gets the first .gguf file in the repository (excludes mmproj files)
-    3. Quantization variant: find a single file ending with the variant name (case insensitive)
-    4. Folder name: downloads all .gguf files in the folder that matches the variant name (case insensitive)
-
-    Examples:
-    - "ggml-org/gpt-oss-120b-GGUF:*" -> downloads all .gguf files in repo
-    - "unsloth/Qwen3-8B-GGUF:qwen3.gguf" -> downloads "qwen3.gguf"
-    - "unsloth/Qwen3-30B-A3B-GGUF" -> downloads "Qwen3-30B-A3B-GGUF.gguf"
-    - "unsloth/Qwen3-8B-GGUF:Q4_1" -> downloads "Qwen3-8B-GGUF-Q4_1.gguf"
-    - "unsloth/Qwen3-30B-A3B-GGUF:Q4_0" -> downloads all files in "Q4_0/" folder
-    )";
-
-    GGUFFiles result;
-    std::vector<std::string> sharded_files;
-    std::string variant_name;
-
-    // (case 0) Wildcard, download everything
-    if (!variant.empty() && variant == "*") {
-        for (const auto& f : repo_files) {
-            if (ends_with_ignore_case(f, ".gguf")) {
-                sharded_files.push_back(f);
-            }
-        }
-        
-        if (sharded_files.empty()) {
-            throw std::runtime_error("No .gguf files found in repository " + checkpoint + ". " + hint);
-        }
-        
-        // Sort to ensure consistent ordering
-        std::sort(sharded_files.begin(), sharded_files.end());
-        
-        // Use first file as primary (this is how llamacpp handles it)
-        variant_name = sharded_files[0];
-    }
-    // (case 1) If variant ends in .gguf or .bin, use it directly
-    else if (!variant.empty() && (ends_with_ignore_case(variant, ".gguf") || ends_with_ignore_case(variant, ".bin"))) {
-        variant_name = variant;
-
-        // Validate file exists in repo
-        bool found = false;
-        for (const auto& f : repo_files) {
-            if (f == variant) {
-                found = true;
-                break;
-            }
-        }
-
-        if (!found) {
-            throw std::runtime_error(
-                "File " + variant + " not found in Hugging Face repository " + checkpoint + ". " + hint
-            );
-        }
-    }
-    // (case 2) If no variant is provided, get the first .gguf file in the repository
-    else if (variant.empty()) {
-        std::vector<std::string> all_variants;
-        for (const auto& f : repo_files) {
-            if (ends_with_ignore_case(f, ".gguf") && !contains_ignore_case(f, "mmproj")) {
-                all_variants.push_back(f);
-            }
-        }
-        
-        if (all_variants.empty()) {
-            throw std::runtime_error(
-                "No .gguf files found in Hugging Face repository " + checkpoint + ". " + hint
-            );
-        }
-        
-        variant_name = all_variants[0];
-    }
-    else {
-        // (case 3) Find a single file ending with the variant name (case insensitive)
-        std::vector<std::string> end_with_variant;
-        std::string variant_suffix = variant + ".gguf";
-        
-        for (const auto& f : repo_files) {
-            if (ends_with_ignore_case(f, variant_suffix) && !contains_ignore_case(f, "mmproj")) {
-                end_with_variant.push_back(f);
-            }
-        }
-        
-        if (end_with_variant.size() == 1) {
-            variant_name = end_with_variant[0];
-        }
-        else if (end_with_variant.size() > 1) {
-            throw std::runtime_error(
-                "Multiple .gguf files found for variant " + variant + ", but only one is allowed. " + hint
-            );
-        }
-        // (case 4) Check whether the variant corresponds to a folder with sharded files (case insensitive)
-        else {
-            std::string folder_prefix = variant + "/";
-            for (const auto& f : repo_files) {
-                if (ends_with_ignore_case(f, ".gguf") && starts_with_ignore_case(f, folder_prefix)) {
-                    sharded_files.push_back(f);
-                }
-            }
-            
-            if (sharded_files.empty()) {
-                throw std::runtime_error(
-                    "No .gguf files found for variant " + variant + ". " + hint
-                );
-            }
-            
-            // Sort to ensure consistent ordering
-            std::sort(sharded_files.begin(), sharded_files.end());
-            
-            // Use first file as primary (this is how llamacpp handles it)
-            variant_name = sharded_files[0];
-        }
-    }
-    
-    result.core_files["variant"] = variant_name;
-    result.sharded_files = sharded_files;
-    
-    // If there is a mmproj file, add it to the core files
-    if (!mmproj.empty()) {
-        bool found = false;
-        for (const auto& f : repo_files) {
-            if (f == mmproj) {
-                found = true;
-                break;
-            }
-        }
-        
-        if (!found) {
-            throw std::runtime_error(
-                "The provided mmproj file " + mmproj + " was not found in " + checkpoint + "."
-            );
-        }
-        
-        result.core_files["mmproj"] = mmproj;
-    }
-    
-    return result;
-}
 
 ModelManager::ModelManager() {
     server_models_ = load_server_models();
@@ -1074,698 +920,6 @@ bool ModelManager::is_model_downloaded(const std::string& model_name,
     // This overload is no longer needed with unified cache, but keep for compatibility
     // Just delegate to the simpler version
     return is_model_downloaded(model_name);
-}
-
-void ModelManager::download_model(const std::string& model_name,
-                                 const std::string& checkpoint,
-                                 const std::string& recipe,
-                                 bool reasoning,
-                                 bool vision,
-                                 bool embedding,
-                                 bool reranking,
-                                 const std::string& mmproj,
-                                 bool do_not_upgrade,
-                                 DownloadProgressCallback progress_callback) {
-    
-    std::string actual_checkpoint = checkpoint;
-    std::string actual_recipe = recipe;
-    std::string actual_mmproj = mmproj;
-    
-    // Check if model exists in registry
-    bool model_registered = model_exists(model_name);
-    
-    if (!model_registered) {
-        // Model not in registry - this must be a user model registration
-        // Validate it has the "user." prefix
-        if (model_name.substr(0, 5) != "user.") {
-            throw std::runtime_error(
-                "When registering a new model, the model name must include the "
-                "`user` namespace, for example `user.Phi-4-Mini-GGUF`. Received: " + 
-                model_name
-            );
-        }
-        
-        // Check that required arguments are provided
-        if (actual_checkpoint.empty() || actual_recipe.empty()) {
-            throw std::runtime_error(
-                "Model " + model_name + " is not registered with Lemonade Server. "
-                "To register and install it, provide the `checkpoint` and `recipe` "
-                "arguments, as well as the optional `reasoning` and `mmproj` arguments "
-                "as appropriate."
-            );
-        }
-        
-        // Validate GGUF models (llamacpp recipe) require a variant
-        if (actual_recipe == "llamacpp") {
-            std::string checkpoint_lower = actual_checkpoint;
-            std::transform(checkpoint_lower.begin(), checkpoint_lower.end(), 
-                          checkpoint_lower.begin(), ::tolower);
-            if (checkpoint_lower.find("gguf") != std::string::npos && 
-                actual_checkpoint.find(':') == std::string::npos) {
-                throw std::runtime_error(
-                    "You are required to provide a 'variant' in the checkpoint field when "
-                    "registering a GGUF model. The variant is provided as CHECKPOINT:VARIANT. "
-                    "For example: Qwen/Qwen2.5-Coder-3B-Instruct-GGUF:Q4_0 or "
-                    "Qwen/Qwen2.5-Coder-3B-Instruct-GGUF:qwen2.5-coder-3b-instruct-q4_0.gguf"
-                );
-            }
-        }
-        
-        std::cout << "Registering new user model: " << model_name << std::endl;
-    } else {
-        // Model is registered - if checkpoint not provided, look up from registry
-        if (actual_checkpoint.empty()) {
-            auto info = get_model_info(model_name);
-            actual_checkpoint = info.checkpoint;
-            actual_recipe = info.recipe;
-        }
-        
-        // Also look up mmproj if not provided (for vision models)
-        if (actual_mmproj.empty()) {
-            auto info = get_model_info(model_name);
-            actual_mmproj = info.mmproj;
-            if (!actual_mmproj.empty()) {
-                std::cout << "[ModelManager] Found mmproj for vision model: " << actual_mmproj << std::endl;
-            }
-        }
-    }
-    
-    // Parse checkpoint
-    std::string repo_id = actual_checkpoint;
-    std::string variant = "";
-    
-    size_t colon_pos = actual_checkpoint.find(':');
-    if (colon_pos != std::string::npos) {
-        repo_id = actual_checkpoint.substr(0, colon_pos);
-        variant = actual_checkpoint.substr(colon_pos + 1);
-    }
-    
-    std::cout << "Downloading model: " << repo_id;
-    if (!variant.empty()) {
-        std::cout << " (variant: " << variant << ")";
-    }
-    std::cout << std::endl;
-    
-    // Check if offline mode
-    const char* offline_env = std::getenv("LEMONADE_OFFLINE");
-    if (offline_env && std::string(offline_env) == "1") {
-        std::cout << "Offline mode enabled, skipping download" << std::endl;
-        return;
-    }
-    
-    // CRITICAL: If do_not_upgrade=true AND model is already downloaded, skip entirely
-    // This prevents unnecessary HuggingFace API queries when we just want to use cached models
-    // The do_not_upgrade flag means:
-    //   - Load/inference endpoints: Don't check HuggingFace for updates (use cache if available)
-    //   - Pull endpoint: Always check HuggingFace for latest version (do_not_upgrade=false)
-    if (do_not_upgrade && is_model_downloaded(model_name)) {
-        std::cout << "[ModelManager] Model already downloaded and do_not_upgrade=true, using cached version" << std::endl;
-        return;
-    }
-    
-    // Use FLM pull for FLM models, otherwise download from HuggingFace
-    if (actual_recipe == "flm") {
-        download_from_flm(actual_checkpoint, do_not_upgrade, progress_callback);
-    } else if (actual_recipe == "llamacpp" || actual_recipe == "whispercpp") {
-        // For llamacpp (GGUF) and whispercpp (.bin) models, use variant-aware download
-        download_from_huggingface(repo_id, variant, actual_mmproj, progress_callback);
-    } else {
-        // For non-GGUF models (oga-*, etc.), download all files (no variant filtering)
-        download_from_huggingface(repo_id, "", "", progress_callback);
-    }
-    
-    // Register if needed
-    if (model_name.substr(0, 5) == "user." || !checkpoint.empty()) {
-        register_user_model(model_name, actual_checkpoint, actual_recipe, 
-                          reasoning, vision, embedding, reranking, actual_mmproj);
-    }
-    
-    // Update cache after successful download
-    update_model_in_cache(model_name, true);
-}
-
-// Download model files from HuggingFace
-// =====================================
-// IMPORTANT: This function ALWAYS queries the HuggingFace API to get the repository
-// file list, then downloads any missing files. It does NOT check do_not_upgrade.
-//
-// The caller (download_model) is responsible for checking do_not_upgrade and
-// calling is_model_downloaded() before invoking this function.
-//
-// Download capabilities by backend:
-//   - Lemonade Router (ModelManager): ✅ Downloads non-FLM models from HuggingFace
-//   - FLM backend: ✅ Downloads FLM models via 'flm pull' command
-//   - llama-server backend: ❌ Cannot download (expects GGUF files pre-cached)
-//   - ryzenai-server backend: ❌ Cannot download (expects ONNX files pre-cached)
-void ModelManager::download_from_huggingface(const std::string& repo_id,
-                                            const std::string& variant,
-                                            const std::string& mmproj,
-                                            DownloadProgressCallback progress_callback) {
-    // Get Hugging Face cache directory
-    std::string hf_cache = get_hf_cache_dir();
-    
-    // Create cache directory structure
-    fs::create_directories(hf_cache);
-    
-    std::string cache_dir_name = "models--";
-    for (char c : repo_id) {
-        if (c == '/') {
-            cache_dir_name += "--";
-        } else {
-            cache_dir_name += c;
-        }
-    }
-    
-    std::string model_cache_path = hf_cache + "/" + cache_dir_name;
-    fs::create_directories(model_cache_path);
-    
-    // Get HF token if available
-    std::map<std::string, std::string> headers;
-    const char* hf_token = std::getenv("HF_TOKEN");
-    if (hf_token) {
-        headers["Authorization"] = "Bearer " + std::string(hf_token);
-    }
-    
-    // Query HuggingFace API to get list of all files in the repository
-    // NOTE: This API call happens EVERY time this function is called, regardless of
-    // whether files are cached. The do_not_upgrade check should happen in the caller
-    // (download_model) to avoid this API call when using cached models.
-    std::string api_url = "https://huggingface.co/api/models/" + repo_id;
-    
-    try {
-        std::cout << "[ModelManager] Fetching repository file list from Hugging Face..." << std::endl;
-        auto response = HttpClient::get(api_url, headers);
-        
-        if (response.status_code != 200) {
-            throw std::runtime_error(
-                "Failed to fetch model info from Hugging Face API (status: " + 
-                std::to_string(response.status_code) + ")"
-            );
-        }
-        
-        auto model_info = JsonUtils::parse(response.body);
-        
-        if (!model_info.contains("siblings") || !model_info["siblings"].is_array()) {
-            throw std::runtime_error("Invalid model info response from Hugging Face API");
-        }
-        
-        // Extract commit hash (sha) from the API response
-        std::string commit_hash;
-        if (model_info.contains("sha") && model_info["sha"].is_string()) {
-            commit_hash = model_info["sha"].get<std::string>();
-            std::cout << "[ModelManager] Using commit hash: " << commit_hash << std::endl;
-        } else {
-            // Fallback to "main" if sha is not available
-            commit_hash = "main";
-            std::cout << "[ModelManager] Warning: No commit hash found in API response, using 'main'" << std::endl;
-        }
-        
-        // Create snapshot directory using commit hash
-        std::string snapshot_path = model_cache_path + "/snapshots/" + commit_hash;
-        fs::create_directories(snapshot_path);
-        
-        // Create refs/main file pointing to this commit (matching huggingface_hub behavior)
-        std::string refs_dir = model_cache_path + "/refs";
-        fs::create_directories(refs_dir);
-        std::string refs_main_path = refs_dir + "/main";
-        std::ofstream refs_file(refs_main_path);
-        if (refs_file.is_open()) {
-            refs_file << commit_hash;
-            refs_file.close();
-        }
-        
-        // Extract list of all files in the repository
-        std::vector<std::string> repo_files;
-        for (const auto& file : model_info["siblings"]) {
-            if (file.contains("rfilename")) {
-                repo_files.push_back(file["rfilename"].get<std::string>());
-            }
-        }
-        
-        std::cout << "[ModelManager] Repository contains " << repo_files.size() << " files" << std::endl;
-        
-        std::vector<std::string> files_to_download;
-        
-        // Check if this is a GGUF model (variant provided) or non-GGUF (variant empty)
-        if (!variant.empty() || !mmproj.empty()) {
-            // GGUF model: Use identify_gguf_models to determine which files to download
-            GGUFFiles gguf_files = identify_gguf_models(repo_id, variant, mmproj, repo_files);
-            
-            // Combine core files and sharded files into one list
-            for (const auto& [key, filename] : gguf_files.core_files) {
-                files_to_download.push_back(filename);
-            }
-            for (const auto& filename : gguf_files.sharded_files) {
-                files_to_download.push_back(filename);
-            }
-            
-            // Also download essential config files if they exist
-            std::vector<std::string> config_files = {
-                "config.json",
-                "tokenizer.json",
-                "tokenizer_config.json",
-                "tokenizer.model"
-            };
-            for (const auto& config_file : config_files) {
-                if (std::find(repo_files.begin(), repo_files.end(), config_file) != repo_files.end()) {
-                    if (std::find(files_to_download.begin(), files_to_download.end(), config_file) == files_to_download.end()) {
-                        files_to_download.push_back(config_file);
-                    }
-                }
-            }
-        } else {
-            // Non-GGUF model (ONNX, etc.): Download all files in repository
-            files_to_download = repo_files;
-        }
-        
-        std::cout << "[ModelManager] Identified " << files_to_download.size() << " files to download:" << std::endl;
-        for (const auto& filename : files_to_download) {
-            std::cout << "  - " << filename << std::endl;
-        }
-        
-        // Create download manifest to track incomplete downloads
-        // This allows us to detect partially downloaded models
-        std::string manifest_path = snapshot_path + "/.download_manifest.json";
-        
-        // Fetch file sizes from the tree API (the models API doesn't include sizes)
-        std::map<std::string, size_t> file_sizes;
-        std::string tree_url = "https://huggingface.co/api/models/" + repo_id + "/tree/main";
-        auto tree_response = HttpClient::get(tree_url, headers);
-        
-        if (tree_response.status_code == 200) {
-            auto tree_info = JsonUtils::parse(tree_response.body);
-            if (tree_info.is_array()) {
-                for (const auto& file : tree_info) {
-                    if (file.contains("path") && file.contains("size")) {
-                        std::string fpath = file["path"].get<std::string>();
-                        size_t fsize = file["size"].get<size_t>();
-                        file_sizes[fpath] = fsize;
-                    }
-                }
-            }
-            std::cout << "[ModelManager] Retrieved file sizes for " << file_sizes.size() << " files" << std::endl;
-        } else {
-            std::cout << "[ModelManager] Warning: Could not fetch file sizes (tree API returned " 
-                      << tree_response.status_code << ")" << std::endl;
-        }
-        
-        // Create manifest with expected files
-        json manifest;
-        manifest["repo_id"] = repo_id;
-        manifest["commit_hash"] = commit_hash;
-        manifest["files"] = json::array();
-        for (const auto& filename : files_to_download) {
-            json file_entry;
-            file_entry["name"] = filename;
-            file_entry["size"] = file_sizes.count(filename) ? file_sizes[filename] : 0;
-            manifest["files"].push_back(file_entry);
-        }
-        
-        // Write manifest (indicates download in progress)
-        JsonUtils::save_to_file(manifest, manifest_path);
-        std::cout << "[ModelManager] Created download manifest" << std::endl;
-        
-        // Download each file with robust retry and resume support
-        int file_index = 0;
-        int total_files = static_cast<int>(files_to_download.size());
-        
-        for (const auto& filename : files_to_download) {
-            file_index++;
-            std::string file_url = "https://huggingface.co/" + repo_id + 
-                                 "/resolve/main/" + filename;
-            std::string output_path = snapshot_path + "/" + filename;
-            
-            // Create parent directory for file (handles folders in filenames)
-            fs::create_directories(fs::path(output_path).parent_path());
-            
-            std::cout << "[ModelManager] Downloading: " << filename << "..." << std::endl;
-            
-            // Send progress update if callback provided (and check for cancellation)
-            if (progress_callback) {
-                DownloadProgress progress;
-                progress.file = filename;
-                progress.file_index = file_index;
-                progress.total_files = total_files;
-                progress.bytes_downloaded = 0;
-                progress.bytes_total = file_sizes.count(filename) ? file_sizes[filename] : 0;
-                progress.percent = 0;
-                if (!progress_callback(progress)) {
-                    std::cout << "[ModelManager] Download cancelled by client" << std::endl;
-                    throw std::runtime_error("Download cancelled");
-                }
-            }
-            
-            utils::DownloadOptions download_opts;
-            download_opts.max_retries = 10;
-            download_opts.initial_retry_delay_ms = 2000;
-            download_opts.max_retry_delay_ms = 120000;
-            download_opts.resume_partial = true;
-            download_opts.low_speed_limit = 1000;
-            download_opts.low_speed_time = 60;
-            download_opts.connect_timeout = 60;
-            
-            // Create progress callback that reports to both console and SSE callback
-            // Returns bool: true = continue, false = cancel
-            utils::ProgressCallback http_progress_cb;
-            if (progress_callback) {
-                http_progress_cb = [&](size_t downloaded, size_t total) -> bool {
-                    DownloadProgress progress;
-                    progress.file = filename;
-                    progress.file_index = file_index;
-                    progress.total_files = total_files;
-                    progress.bytes_downloaded = downloaded;
-                    progress.bytes_total = total;
-                    progress.percent = (total > 0) ? static_cast<int>((downloaded * 100) / total) : 0;
-                    return progress_callback(progress);  // Propagate cancellation
-                };
-            } else {
-                // Default console progress callback (never cancels)
-                http_progress_cb = utils::create_throttled_progress_callback();
-            }
-            
-            auto result = HttpClient::download_file(
-                file_url,
-                output_path,
-                http_progress_cb,
-                headers,
-                download_opts
-            );
-            
-            // Check if download was cancelled
-            if (result.cancelled) {
-                std::cout << "[ModelManager] Download cancelled by client" << std::endl;
-                throw std::runtime_error("Download cancelled");
-            }
-            
-            if (result.success) {
-                std::cout << "\n[ModelManager] Downloaded: " << filename << std::endl;
-            } else {
-                // Build a detailed error message
-                std::ostringstream error_msg;
-                error_msg << "Failed to download file: " << filename << "\n";
-                error_msg << "URL: " << file_url << "\n";
-                error_msg << result.error_message;
-                
-                // If there's a partial file, provide helpful information
-                if (fs::exists(output_path)) {
-                    size_t partial_size = fs::file_size(output_path);
-                    if (partial_size > 0) {
-                        error_msg << "\n\n[INFO] Partial download preserved at: " << output_path;
-                        error_msg << "\n[INFO] Partial size: " << std::fixed << std::setprecision(1) 
-                                  << (partial_size / (1024.0 * 1024.0)) << " MB";
-                        error_msg << "\n[INFO] Run the command again to resume from where it left off.";
-                    }
-                }
-                
-                throw std::runtime_error(error_msg.str());
-            }
-        }
-        
-        // Validate all expected files exist after download
-        std::cout << "[ModelManager] Validating downloaded files..." << std::endl;
-        bool all_valid = true;
-        for (const auto& filename : files_to_download) {
-            std::string expected_path = snapshot_path + "/" + filename;
-            std::string partial_path = expected_path + ".partial";
-            
-            // Check for .partial file (incomplete download)
-            if (fs::exists(partial_path)) {
-                all_valid = false;
-                std::cerr << "[ModelManager] Incomplete file found: " << filename << ".partial" << std::endl;
-                continue;
-            }
-            
-            if (!fs::exists(expected_path)) {
-                all_valid = false;
-                std::cerr << "[ModelManager] Missing file: " << filename << std::endl;
-                continue;
-            }
-            
-            // Verify file size if we have expected size
-            if (file_sizes.count(filename) && file_sizes[filename] > 0) {
-                size_t actual_size = fs::file_size(expected_path);
-                size_t expected_size = file_sizes[filename];
-                if (actual_size != expected_size) {
-                    all_valid = false;
-                    std::cerr << "[ModelManager] Size mismatch for " << filename 
-                              << ": expected " << expected_size << " bytes, got " << actual_size << " bytes" << std::endl;
-                }
-            }
-        }
-        
-        if (!all_valid) {
-            throw std::runtime_error(
-                "Download validation failed. Some files are incomplete or missing. "
-                "Run the command again to resume."
-            );
-        }
-        
-        // All files validated - remove manifest to mark download as complete
-        if (fs::exists(manifest_path)) {
-            fs::remove(manifest_path);
-            std::cout << "[ModelManager] Removed download manifest (download complete)" << std::endl;
-        }
-        
-        // Send completion event
-        // Note: We ignore the return value here since the download is already complete
-        // If the client disconnected, the write will simply fail silently
-        if (progress_callback) {
-            DownloadProgress progress;
-            progress.complete = true;
-            progress.file_index = total_files;
-            progress.total_files = total_files;
-            progress.percent = 100;
-            (void)progress_callback(progress);  // Ignore return - download already complete
-        }
-        
-        std::cout << "[ModelManager] ✓ All files downloaded and validated successfully!" << std::endl;
-        std::cout << "[ModelManager DEBUG] Download location: " << snapshot_path << std::endl;
-        
-    } catch (const std::exception& e) {
-        // Don't log here - let the caller (Server) handle error logging
-        throw;
-    }
-}
-
-void ModelManager::download_from_flm(const std::string& checkpoint, 
-                                     bool do_not_upgrade,
-                                     DownloadProgressCallback progress_callback) {
-    std::cout << "[ModelManager] Pulling FLM model: " << checkpoint << std::endl;
-    
-    // Ensure FLM is installed
-    std::cout << "[ModelManager] Checking FLM installation..." << std::endl;
-    backends::FastFlowLMServer flm_installer("info", this);
-    try {
-        flm_installer.install();
-    } catch (const std::exception& e) {
-        std::cerr << "[ModelManager ERROR] FLM installation failed: " << e.what() << std::endl;
-        throw;
-    }
-    
-    // Find flm executable
-    std::string flm_path = "flm";
-    
-    // Prepare arguments
-    std::vector<std::string> args = {"pull", checkpoint};
-    if (!do_not_upgrade) {
-        args.push_back("--force");
-    }
-    
-    std::cout << "[ProcessManager] Starting process: \"" << flm_path << "\"";
-    for (const auto& arg : args) {
-        std::cout << " \"" << arg << "\"";
-    }
-    std::cout << std::endl;
-    
-    // State for parsing FLM output
-    int total_files = 0;
-    int current_file_index = 0;
-    std::string current_filename;
-    bool cancelled = false;
-    
-    // Run flm pull command and parse output
-    int exit_code = utils::ProcessManager::run_process_with_output(
-        flm_path, args,
-        [&](const std::string& line) -> bool {
-            // Always print the line to console
-            std::cout << line << std::endl;
-            
-            // Parse FLM output to extract progress information
-            // Pattern: "[FLM]  Downloading X/Y: filename"
-            if (line.find("[FLM]  Downloading ") != std::string::npos && 
-                line.find("/") != std::string::npos && 
-                line.find(":") != std::string::npos) {
-                
-                // Extract "X/Y: filename" from "[FLM]  Downloading X/Y: filename"
-                size_t start = line.find("Downloading ") + 12;
-                size_t slash = line.find("/", start);
-                size_t colon = line.find(":", slash);
-                
-                if (slash != std::string::npos && colon != std::string::npos) {
-                    try {
-                        current_file_index = std::stoi(line.substr(start, slash - start));
-                        total_files = std::stoi(line.substr(slash + 1, colon - slash - 1));
-                        current_filename = line.substr(colon + 2);  // Skip ": "
-                        
-                        // Send progress update
-                        if (progress_callback) {
-                            DownloadProgress progress;
-                            progress.file = current_filename;
-                            progress.file_index = current_file_index;
-                            progress.total_files = total_files;
-                            progress.bytes_downloaded = 0;
-                            progress.bytes_total = 0;
-                            progress.percent = (total_files > 0) ? 
-                                ((current_file_index - 1) * 100 / total_files) : 0;
-                            
-                            if (!progress_callback(progress)) {
-                                cancelled = true;
-                                return false;  // Kill the process
-                            }
-                        }
-                    } catch (...) {
-                        // Ignore parse errors
-                    }
-                }
-            }
-            // Pattern: "[FLM]  Downloading: XX.X% (XXX.XMB / XXX.XMB)"
-            else if (line.find("[FLM]  Downloading: ") != std::string::npos && 
-                     line.find("%") != std::string::npos) {
-                
-                // Extract percentage and bytes
-                size_t start = line.find("Downloading: ") + 13;
-                size_t pct_end = line.find("%", start);
-                
-                if (pct_end != std::string::npos) {
-                    try {
-                        std::string pct_str = line.substr(start, pct_end - start);
-                        double file_percent = std::stod(pct_str);
-                        
-                        // Try to extract bytes (XXX.XMB / XXX.XMB)
-                        size_t open_paren = line.find("(", pct_end);
-                        size_t slash = line.find("/", open_paren);
-                        size_t close_paren = line.find(")", slash);
-                        
-                        size_t bytes_downloaded = 0;
-                        size_t bytes_total = 0;
-                        
-                        if (open_paren != std::string::npos && slash != std::string::npos) {
-                            std::string downloaded_str = line.substr(open_paren + 1, slash - open_paren - 1);
-                            std::string total_str = line.substr(slash + 1, close_paren - slash - 1);
-                            
-                            // Parse "XXX.XMB" format
-                            auto parse_size = [](const std::string& s) -> size_t {
-                                double val = 0;
-                                size_t mb_pos = s.find("MB");
-                                size_t gb_pos = s.find("GB");
-                                size_t kb_pos = s.find("KB");
-                                
-                                if (mb_pos != std::string::npos) {
-                                    val = std::stod(s.substr(0, mb_pos));
-                                    return static_cast<size_t>(val * 1024 * 1024);
-                                } else if (gb_pos != std::string::npos) {
-                                    val = std::stod(s.substr(0, gb_pos));
-                                    return static_cast<size_t>(val * 1024 * 1024 * 1024);
-                                } else if (kb_pos != std::string::npos) {
-                                    val = std::stod(s.substr(0, kb_pos));
-                                    return static_cast<size_t>(val * 1024);
-                                }
-                                return 0;
-                            };
-                            
-                            bytes_downloaded = parse_size(downloaded_str);
-                            bytes_total = parse_size(total_str);
-                        }
-                        
-                        // Send progress update with byte-level info
-                        if (progress_callback) {
-                            DownloadProgress progress;
-                            progress.file = current_filename;
-                            progress.file_index = current_file_index;
-                            progress.total_files = total_files;
-                            progress.bytes_downloaded = bytes_downloaded;
-                            progress.bytes_total = bytes_total;
-                            // Use intra-file percent when we have byte-level progress
-                            progress.percent = static_cast<int>(file_percent);
-                            
-                            if (!progress_callback(progress)) {
-                                cancelled = true;
-                                return false;  // Kill the process
-                            }
-                        }
-                    } catch (...) {
-                        // Ignore parse errors
-                    }
-                }
-            }
-            // Pattern: "[FLM]  Overall progress: XX.X% (X/Y files)"
-            else if (line.find("[FLM]  Overall progress: ") != std::string::npos) {
-                size_t start = line.find("progress: ") + 10;
-                size_t pct_end = line.find("%", start);
-                
-                if (pct_end != std::string::npos) {
-                    try {
-                        int overall_percent = static_cast<int>(std::stod(line.substr(start, pct_end - start)));
-                        
-                        if (progress_callback) {
-                            DownloadProgress progress;
-                            progress.file = current_filename;
-                            progress.file_index = current_file_index;
-                            progress.total_files = total_files;
-                            progress.bytes_downloaded = 0;  // Not available for overall progress
-                            progress.bytes_total = 0;
-                            progress.percent = overall_percent;
-                            
-                            if (!progress_callback(progress)) {
-                                cancelled = true;
-                                return false;  // Kill the process
-                            }
-                        }
-                    } catch (...) {
-                        // Ignore parse errors
-                    }
-                }
-            }
-            // Pattern: "[FLM]  Missing files (N):"
-            else if (line.find("[FLM]  Missing files (") != std::string::npos) {
-                size_t start = line.find("(") + 1;
-                size_t end = line.find(")", start);
-                if (end != std::string::npos) {
-                    try {
-                        total_files = std::stoi(line.substr(start, end - start));
-                    } catch (...) {
-                        // Ignore parse errors
-                    }
-                }
-            }
-            
-            return true;  // Continue
-        },
-        "",  // Working directory
-        3600  // 1 hour timeout for large model downloads
-    );
-    
-    if (cancelled) {
-        std::cout << "[ModelManager] FLM download cancelled by client" << std::endl;
-        throw std::runtime_error("Download cancelled");
-    }
-    
-    if (exit_code != 0) {
-        std::cerr << "[ModelManager ERROR] FLM pull failed with exit code: " << exit_code << std::endl;
-        throw std::runtime_error("FLM pull failed with exit code: " + std::to_string(exit_code));
-    }
-    
-    // Send completion event
-    if (progress_callback) {
-        DownloadProgress progress;
-        progress.complete = true;
-        progress.file_index = total_files;
-        progress.total_files = total_files;
-        progress.percent = 100;
-        (void)progress_callback(progress);  // Ignore return - download already complete
-    }
-    
-    std::cout << "[ModelManager] FLM model pull completed successfully" << std::endl;
 }
 
 void ModelManager::delete_model(const std::string& model_name) {
